pyproject.toml
[tool.poetry]
name = "elevatus-evaluation-task"
version = "1.0.0"
description = "This poetry toml file to manage and handle project packages"
authors = ["Hisham Nasrallah <hisham.nasralla@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
annotated-types = "0.6.0"
anyio = "4.2.0"
build = "1.0.3"
cachecontrol = "0.13.1"
certifi = "2023.11.17"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
click = "8.1.7"
crashtest = "0.4.1"
cryptography = "41.0.7"
distlib = "0.3.8"
dnspython = "2.4.2"
dulwich = "0.21.7"
ecdsa = "0.18.0"
email-validator = "2.1.0.post1"
exceptiongroup = "1.2.0"
fastapi = "0.108.0"
fastjsonschema = "2.19.1"
filelock = "3.13.1"
greenlet = "3.0.3"
h11 = "0.14.0"
httptools = "0.6.1"
idna = "3.6"
importlib-metadata = "7.0.1"
installer = "0.7.0"
jaraco-classes = "3.3.0"
jeepney = "0.8.0"
keyring = "24.3.0"
more-itertools = "10.2.0"
motor = "3.3.2"
msgpack = "1.0.7"
packaging = "23.2"
passlib = "1.7.4"
pexpect = "4.9.0"
pkginfo = "1.9.6"
platformdirs = "3.11.0"
poetry = "1.7.1"
poetry-core = "1.8.1"
poetry-plugin-export = "1.6.0"
ptyprocess = "0.7.0"
pyasn1 = "0.5.1"
pycparser = "2.21"
pydantic = "2.5.3"
pydantic-settings = "2.1.0"
pydantic-core = "2.14.6"
pydash = "7.0.6"
pyjwt = "2.8.0"
pymongo = "4.6.1"
pyproject-hooks = "1.0.0"
python-dotenv = "1.0.0"
python-jose = "3.3.0"
pyyaml = "6.0.1"
rapidfuzz = "3.6.1"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
rsa = "4.9"
secretstorage = "3.3.3"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.25"
sqlalchemy-utils = "0.41.1"
starlette = "0.32.0.post1"
tomli = "2.0.1"
tomlkit = "0.12.3"
trove-classifiers = "2024.1.8"
typing-extensions = "4.9.0"
urllib3 = "2.1.0"
uvicorn = "0.25.0"
uvloop = "0.19.0"
virtualenv = "20.25.0"
watchfiles = "0.21.0"
websockets = "12.0"
zipp = "3.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
